moved speakers into their own folder
