+++ src/domains/Speaker/SpeakerCard/SpeakerCard.tsx
