+++ src/domains/Speaker/SpeakerCard/index.tsx
